refactor: 공통 영역에서 세분화
